+++ packages/@twokeys/addons/src/registry.ts
@@ -28,7 +28,7 @@ import sqlite3 from "sqlite3";
 import { v4 as uuidv4 } from "uuid";
 import { promises as fs, constants as fsconstants } from "fs";
 import { join } from "path";
-import { Logger, TwoKeysProperties, CodedError } from "@twokeys/core";
+import { Logger, CodedError } from "@twokeys/core";
 import { DEFAULT_REGISTRY_ROOT_PACKAGE_JSON, REGISTRY_FILE_NAME, CREATE_REGISTRY_DB_QUERY, REGISTRY_TABLE_NAME, REGISTRY_MODULE_FOLDER } from "./util/constants";
 import { Package, PackageInDB, TWOKEYS_ADDON_TYPES_ARRAY, TwokeysPackageInfo, ValidatorReturn, TWOKEYS_ADDON_TYPES, TWOKEYS_ADDON_TYPE_EXECUTOR, TWOKEYS_ADDON_TYPE_DETECTOR, TWOKEYS_ADDON_TYPE_SINGLE } from "./util/interfaces";
 import { AddOnModulesCollection, TaskFunction, BaseAddon } from "./module-interfaces";
