+++ packages/@twokeys/server/src/routes/api.ts
@@ -30,26 +30,18 @@ import Logger from "../util/logger";
 import {EvDevValues } from "../util/interfaces";
 import { run_hotkey, fetch_hotkey } from "../util/ahk";
 import { CONFIG_FILE } from "../util/constants";
-import { BaseHotkeyTypeSingle, DetectorConfig, Hotkey, HotkeyTypeKeypressValue, HotkeyTypeSingle, Keyboard, ProjectConfig } from "@twokeys/core/lib/interfaces";
+import { ProjectConfig } from "@twokeys/core/lib/interfaces";
 import { loadMainConfig } from "@twokeys/core/lib/config";
 import { TWOKEYS_MAIN_CONFIG_DEFAULT_PATH } from "@twokeys/core/lib/constants";
-import { AddOnsRegistry, ExecutorExecConfig } from "@twokeys/addons";
+import { AddOnsRegistry } from "@twokeys/addons";
 import { loadDetectors } from "./loadDetectors";
 import { loadExecutors } from "./loadExecutors";
-import { CodedError } from "@twokeys/core";
-import { ERR_BAD_EVENT_TYPE } from "../util/errCodes";
+import getTriggerHotkey from "./triggerHotkey";
 
 export const logger: Logger = new Logger({
 	name: "api",
 });
 const router = Router();
-
-/**
- * Helper function to check if obeject has property
- */
-function hasOwnProperty<T extends { [key: string]: any }, K extends keyof T>(obj: T, prop: K): boolean  {
-	return Object.prototype.hasOwnProperty.call(obj, prop);
-}
 
 /**
  * Needed so we can use async/await
@@ -127,130 +119,7 @@ export default async function getAPI(projectConfig: ProjectConfig, projectDir: s
 	 * }
 	 * ```
 	 */
-	router.post("/post/:detector/:keyboard/trigger", (req, res, next) => {
-		const { detector: detectorName, keyboard: keyboardName } = req.params;
-		logger.info(`Got trigger for detector ${detectorName}, keyboard ${keyboardName}`);
-
-		logger.info("Validating POST body");
-		if (!hasOwnProperty(req, "body") || typeof req.body !== "object" || typeof req.body.hotkey !== "string") {
-			logger.err("Invalid POST body! Properties were missing!");
-			res.statusCode = 422;
-			res.json({
-				message: "Invalid POST body! Properties were missing!",
-			});
-			return;
-		}
-
-		const hotkey = req.body.hotkey as string;
-		const eventType: keyof HotkeyTypeKeypressValue = req.body.event || "down";
-		if (eventType !== "up" && eventType !== "down" && eventType !== "hold") {
-			logger.err("Bad event field given!");
-			res.statusCode = 422;
-			res.json({
-				message: "Bad event field given!",
-			});
-			return;
-		}
-
-		// 1: Grab config
-		logger.debug(`Grabbing config for detector ${detectorName}, keyboard ${keyboardName}...`);
-		if (!detectors.has(detectorName)) {
-			logger.err(`Detector ${detectorName} not found!`);
-			res.statusCode = 404;
-			res.json({
-				message: "Detector Not Found"
-			});
-			return;
-		}
-		const detector = detectors.get(detectorName) as DetectorConfig;
-		if (hasOwnProperty(detector.keyboards, keyboardName)) {
-			logger.debug(`Keybaord ${keyboardName} found`)
-			const keyboard = detector.keyboards[keyboardName];
-
-			if (!hasOwnProperty(keyboard.hotkeys, hotkey)) {
-				logger.err(`Hotkey ${hotkey} not found!`);
-				res.statusCode = 404;
-				res.json({
-					message: "Hotkey Not Found"
-				});
-				return;
-			}
-
-			// eslint-disable-next-line no-inner-declarations
-			function isMultiHotkey(hotkey: Hotkey): hotkey is HotkeyTypeKeypressValue {
-				if (hasOwnProperty(hotkey, "up") || hasOwnProperty(hotkey, "down") || hasOwnProperty(hotkey, "hold")) {
-					return true;
-				} else {
-					return false;
-				}
-			}
-
-			// eslint-disable-next-line no-inner-declarations
-			async function executeHotKey(hotkey: HotkeyTypeSingle, hotkeyCode: string, keyboard: Keyboard): Promise<void> {
-				logger.info(`Executing hotkey ${hotkey}...`);
-				const executorToCall = hotkey.executor || keyboard.executors.default;
-				const configForExecutor: ExecutorExecConfig<{ [key: string]: any }> = {
-					hotkey: {
-						...(keyboard.executors[executorToCall] || {}), // Falback in case no config
-						...hotkey,
-					},
-					hotkeyCode,
-					executorDefaultConfig: (keyboard.executors[executorToCall] || {}),
-					keyboard,
-				};
-				logger.debug(`Providing ${JSON.stringify(configForExecutor)} to executor`);
-				if (!hasOwnProperty(executors, executorToCall)) {
-					logger.err(`Executor ${executorToCall} not found installed!`);
-					throw new CodedError("Executor to use not found!");
-				}
-				await executors[executorToCall].call(executors[executorToCall].execute, configForExecutor);
-				return;
-			}
-
-			const theHotkey = keyboard.hotkeys[hotkey];
-			if (isMultiHotkey(theHotkey)) {
-				if (typeof theHotkey[eventType] !== "object") {
-					return next(new CodedError(`Hotkey event ${eventType} not found!`, ERR_BAD_EVENT_TYPE));
-				} else {
-					executeHotKey(theHotkey[eventType] as HotkeyTypeSingle, hotkey, keyboard)
-						.catch(next)
-						.then(() => {
-							logger.info("Execution done.");
-						});
-					// Send back to prevent timeout from long hotkeys
-					res.statusCode = 200;
-					res.json({
-						message: "Run triggered",
-					});
-					res.end();
-					return;
-				}
-			} else {
-				// TODO: Execute on a different thread, because the server hangs and fails any in progress runs if it is still waiting for this
-				executeHotKey(theHotkey, hotkey, keyboard)
-					.catch(next)
-					.then(() => {
-						logger.info("Execution done.");
-					});
-				// Send back to prevent timeout from long hotkeys
-				res.statusCode = 200;
-				res.json({
-					message: "Run triggered",
-				});
-				res.end();
-				return;
-			}
-
-		} else {
-			logger.err(`Keyboard ${keyboardName} not found!`);
-			res.statusCode = 404;
-			res.json({
-				message: "Keyboard Not Found"
-			});
-			return;
-		}
-
-	});
+	router.post("/post/:detector/:keyboard/trigger", getTriggerHotkey(detectors, executors));
 
 	/**
 	 * Trigger a hotkey
